[DOC] fix description of cookie "insert" and "indirect" modes
The doc was wrong as the insert mode by default does not insert in
direct requests, and by default transmits the cookies to the server.
This was right in the old doc and it has not changed since the
beginning.
(cherry picked from commit a79094d3e41e9f8e64b709c5dd1cb5bf6fdae93d)