[BUG] http: use correct ACL pointer when evaluating authentication
req_acl was used instead of req_acl_final. As a matter of luck, both
happen to be the same at this point, but this is not granted in the
future.

This fix should be backported to 1.4.
(cherry picked from commit 910ef306bc98f061cfa74c253b5714f6161b6f7b)