BUG/MEDIUM: don't trim last spaces from headers consisting only of spaces
Commit bc044ac fixed header parsing so that trailing spaces were not part
of the returned string. Unfortunately, if a header only had spaces, the
last spaces were trimmed past the beginning of the value, causing a negative
length to be returned.

A quick code review shows that there should be no impact since the only
places where the vlen is used are either compared to a specific value or
with explicit contents (eg: digits).

This must be backported to 1.4.
(cherry picked from commit 275600b6c7d5308d6c43caa8f9451bf944204437)